+++ code/src/cljs/sixsq/nuvla/ui/utils/semantic_ui.cljs
@@ -36,7 +36,7 @@
 (def Button (r/adapt-react-class semantic/Button))
 (def ButtonGroup (r/adapt-react-class semantic/ButtonGroup))
 (def ButtonContent (r/adapt-react-class semantic/ButtonContent))
-(def ButtonOr (r/adapt-react-class semantic/ButtonOr))
+;(def ButtonOr (r/adapt-react-class semantic/ButtonOr))
 
 (def Card (r/adapt-react-class semantic/Card))
 (def CardContent (r/adapt-react-class semantic/CardContent))
@@ -46,7 +46,9 @@
 (def CardMeta (r/adapt-react-class semantic/CardMeta))
 
 (def Checkbox (r/adapt-react-class semantic/Checkbox))
-
+;(def Comment (r/adapt-react-class semantic/Comment))
+;(def CommentAvatar (r/adapt-react-class semantic/CommentAvatar))
+;(def CommentContent (r/adapt-react-class semantic/CommentContent))
 (def Container (r/adapt-react-class semantic/Container))
 
 (def DatePicker (r/adapt-react-class date-picker/default))
@@ -150,8 +152,8 @@
 
 (def Radio (r/adapt-react-class semantic/Radio))
 (def Responsive (r/adapt-react-class semantic/Responsive))
-(def Reveal (r/adapt-react-class semantic/Reveal))
-(def RevealContent (r/adapt-react-class semantic/RevealContent))
+;(def Reveal (r/adapt-react-class semantic/Reveal))
+;(def RevealContent (r/adapt-react-class semantic/RevealContent))
 
 (def Segment (r/adapt-react-class semantic/Segment))
 (def SegmentGroup (r/adapt-react-class semantic/SegmentGroup))